+++ src/repositories/index.ts
@@ -1,8 +1,8 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-export * from './StyleguideFolderRepository';
-export * from './StyleguideLibraryRepository';
-export * from './StyleguideLinkRepository';
-export * from './StyleguideNavigationItemRepository';
-export * from './StyleguideNavigationRepository';
-export * from './StyleguidePageRepository';
+export * from './GuidelineFolderRepository';
+export * from './GuidelineLibraryRepository';
+export * from './GuidelineLinkRepository';
+export * from './GuidelineNavigationItemRepository';
+export * from './GuidelineNavigationRepository';
+export * from './GuidelinePageRepository';
